Refactoring: Move ContactsVocabulary into vocabularies.py.
